requirements.txt
asgiref==3.7.2
Django==4.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.18
django-widget-tweaks==1.4.12
gunicorn==20.1.0
Pillow==10.0.0
psycopg2==2.9.6
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.7.0
